requirements.txt
# Temporal Python SDK (choose the latest 1.x that exists on PyPI)
temporalio>=1.12.0,<2.0

# Playwright for web scraping
playwright>=1.36.0

# Discord bot framework
discord.py>=2.5.2